moved send invite to own component, refactor css
