+++ src/components/InputComponents.js
@@ -1,13 +1,9 @@
-import { React, useState } from "react";
 import styled from "styled-components";
 import { Field, useField } from "formik";
-import TextField from "@material-ui/core/TextField";
-import Autocomplete from "@material-ui/lab/Autocomplete";
+import { TextField } from "formik-material-ui";
 import MenuItem from "@material-ui/core/MenuItem";
 import { TimePicker, DatePicker } from "formik-material-ui-pickers";
-import { useQuery } from "@apollo/client";
-import { StyledButton } from "./ComponentStyles";
-import { ALL_USERS } from "../queries";
+import { InputRow } from "./ComponentStyles"
 
 export const TextInput = ({ ...props }) => {
   return (
@@ -39,7 +35,6 @@ export const PickDate = ({ ...props }) => {
 };
 
 export const DropDown = ({ name, options, ...props }) => {
-  console.log(options);
   return (
     <InputRow>
       <Field name={name} {...props} component={TextField} type="text" select>
@@ -103,42 +98,6 @@ export const FormTextArea = ({ label, canEdit, ...props }) => {
   );
 };
 
-export const InvitePlayers = () => {
-  const allUsers = useQuery(ALL_USERS);
-  const players = allUsers?.data?.users?.edges.map((user) => ({
-    name: user.node.name,
-  }));
-  const [newName, setNewName] = useState("");
-
-  const onOptionChange = (value) => {
-    setNewName(value);
-  };
-
-  const handleInvite = (event) => {
-    event.preventDefault();
-    //todo send invite
-    console.log(newName);
-    setNewName("");
-  };
-
-  if (allUsers.loading) {
-    return <label>Ladataan</label>;
-  }
-  return (
-    <InputRow>
-      <Autocomplete
-        onChange={(event, value) => onOptionChange(value)}
-        options={players}
-        getOptionSelected={(players) => players.name}
-        getOptionLabel={(players) => players.name}
-        loadingText={"Ladataan..."}
-        renderInput={(params) => <TextField {...params} />}
-      />
-      <StyledButton onClick={handleInvite}>Kutsu</StyledButton>
-    </InputRow>
-  );
-};
-
 const RadioContainer = styled.div`
   width: 100%;
   margin-top: 1rem;
@@ -180,75 +139,3 @@ const RadioContainer = styled.div`
   }
 `;
 
-//override material-ui css
-const InputRow = styled.div`
-  display: flex;
-  margin-bottom: 1rem;
-
-  label {
-    font-size: ${(props) => props.theme.fontSizes.medium};
-    flex: 1;
-    color: white !important;
-  }
-
-  .MuiFormControl-root {
-    flex-direction: row;
-    margin: 0;
-  }
-
-  .MuiInputLabel-formControl {
-    transform: none;
-    position: relative;
-  }
-
-  .MuiInput-root,
-  .MuiTextField-root {
-    flex: 1;
-  }
-
-  .MuiInput-underline::before {
-    transition: none;
-    border-bottom: 1px solid white;
-  }
-
-  .MuiInput-underline::after {
-    transition: none;
-    border-bottom: none;
-    border-color: white;
-  }
-
-  .MuiInputBase-input {
-    color: white;
-    text-align: center;
-  }
-
-  & .MuiFormLabel-root {
-    display: flex;
-    align-items: flex-end;
-    font-family: inherit;
-  }
-
-  /* .MuiSelect-nativeInput {
-    display: none;
-  } */
-
-  .MuiSelect-select.MuiSelect-select {
-    padding-right: 0;
-  }
-
-  .MuiSvgIcon-root {
-    color: white;
-  }
-
-  .MuiFormHelperText-root {
-    position: absolute;
-    left: 0;
-  }
-
-  .MuiAutocomplete-root {
-    width: 100%;
-    .MuiSvgIcon-root {
-      display: none;
-    }
-  }
-`;
